+++ LyricsEverywhere/app/src/main/java/com/aleksejantonov/lyricseverywhere/ui/base/BaseData.kt
@@ -3,7 +3,7 @@ package com.aleksejantonov.lyricseverywhere.ui.base
 //Маркер для слияния списков
 interface BaseData : ListItem {
 
-  var artistId: Int
+  var artistId: Long
 
   override fun equals(other: Any?): Boolean
 }
